Use the latest version of RequireJS
Updated the package JSON to use the latest version of RequireJS (I need
features that were introduced in 2.1.10).

Using the semver ~ to ensure that patch releases are pulled in
automatically instead of being stuck on a specific version.

This has been tested on a significantly complex project without any
problems.